Reintroduced debug line drawing and Raycasting (2D)
- Debug::DrawLine now works as expected, although currently we cannot change the colour of this line yet. This will require slight changes to how shader data is handled
- Raycasting reintroduced. Still doesn't use a quad tree style system, might be worth doing in the future to reduce the number of colliders considered. Currently considers all colliders within the level

+++ ChaosEngine/src/Chaos/DataTypes/ChaoticArray.h
@@ -74,7 +74,7 @@ namespace Chaos
 				return false;
 			}
 			Data[index] = nullptr;
-			FreeSlots[FreeSlotsSize] = index;
+			FreeSlots[FreeSlotsSize] = (int)index;
 			++FreeSlotsSize;
 			return true;
 		}
